support 64bit builds & CEF1 build 1453

+++ CefGlue/Interop/Struct/cef_browser_settings_t.cs
@@ -17,7 +17,7 @@ namespace CefGlue.Interop
         /// <summary>
         /// Size of this structure.
         /// </summary>
-        public int size;
+        public IntPtr size;
 
         ///
         // Disable drag & drop of URLs from other windows.
